pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Felipe", email = "bolas@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}
